+++ examples/resources/hopsworksai_backup/resource.tf
@@ -1,0 +1,5 @@
+
+resource "hopsworksai_backup" "backup" {
+  cluster_id  = "<CLUSTER_ID>"
+  backup_name = "my-test-backup"
+}