+++ test-fixtures/azure/main.tf
@@ -55,3 +55,47 @@ resource "azurerm_ssh_public_key" "key" {
   location            = data.azurerm_resource_group.rg.location
   public_key          = var.ssh_public_key
 }
+
+resource "azurerm_virtual_network" "vnet" {
+  name                = var.virtual_network_name
+  location            = data.azurerm_resource_group.rg.location
+  resource_group_name = data.azurerm_resource_group.rg.name
+  address_space       = ["10.240.0.0/16"]
+}
+
+resource "azurerm_subnet" "subnet" {
+  name                 = "${var.virtual_network_name}-subnet"
+  resource_group_name  = data.azurerm_resource_group.rg.name
+  virtual_network_name = azurerm_virtual_network.vnet.name
+  address_prefixes     = ["10.240.1.0/24"]
+}
+
+resource "azurerm_network_security_group" "security_group" {
+  name                = "${var.virtual_network_name}-security-group"
+  location            = data.azurerm_resource_group.rg.location
+  resource_group_name = data.azurerm_resource_group.rg.name
+
+  security_rule {
+    name                       = "Http"
+    priority                   = 100
+    direction                  = "Inbound"
+    access                     = "Allow"
+    protocol                   = "Tcp"
+    source_port_range          = "*"
+    destination_port_range     = "80"
+    source_address_prefix      = "*"
+    destination_address_prefix = "*"
+  }
+
+  security_rule {
+    name                       = "Https"
+    priority                   = 110
+    direction                  = "Inbound"
+    access                     = "Allow"
+    protocol                   = "Tcp"
+    source_port_range          = "*"
+    destination_port_range     = "443"
+    source_address_prefix      = "*"
+    destination_address_prefix = "*"
+  }
+}